Storing_in_vectorstore/requirements.txt
python-dotenv
pandas
PyMuPDF
transformers
langchain
faiss-cpu
boto3
langchain-aws